Move Proofs out of non-interactive folder
